Move Signed<T> to the key module
